Use objc2 for macOS font smoothing determination
`objc2` / `objc2-foundation` is much more type-safe than `cocoa`, and
automatically ensures that memory management-rules are upheld.

See also https://github.com/alacritty/copypasta/pull/74.

+++ src/ft/fc/char_set.rs
@@ -38,7 +38,7 @@ impl CharSetRef {
     }
 
     pub fn count(&self) -> u32 {
-        unsafe { FcCharSetCount(self.as_ptr()) as u32 }
+        unsafe { FcCharSetCount(self.as_ptr()) }
     }
 
     pub fn union(&self, other: &CharSetRef) -> CharSet {
